requirements.txt
Flask==2.0.1
opencv-contrib-python==4.5.3
opencv-python==4.5.3
Pillow==8.3.2
python-dotenv==0.15.0